generalize the "lazy folder" concept
makes it easy to use lazy intialized folder in other places
